pyproject.toml
[tool.poetry]
name = "speech-transformer-pl"
version = "0.1.0"
description = ""
authors = ["HSJeong5744 <jhs0256@gmail.com>"]
readme = "README.md"
packages = [{include = "speech_transformer_pl"}]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^2.0.3"
sentencepiece = "^0.1.99"
tqdm = "^4.65.0"
hydra-core = "^1.3.2"
lightning = "^2.0.6"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
isort = "^5.12.0"
ipykernel = "^6.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
